% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PushConceptSets.R
\name{conceptPump}
\alias{conceptPump}
\title{Concept pumping function}
\usage{
conceptPump(
  mappedValues = export,
  strings = VSAC_strings[1],
  project = "[Test Concept Pump]",
  conn = connection,
  schema = "",
  mapToStandard = FALSE,
  Descendants = FALSE,
  Exclude = FALSE,
  Atlas = baseUrl,
  authMethod = "db",
  Username = "ATLAS_USER",
  Password = "ATLAS_PASSWORD"
)
}
\arguments{
\item{mappedValues}{A dataframe of strings and concepts}

\item{strings}{A vector of the value set names}

\item{project}{A string to append for the concept set name in ATLAS}

\item{conn}{A valid CDM database}

\item{schema}{A valid CDM schema}

\item{mapToStandard}{A flag for if the concepts require mapping to standard values}

\item{Descendants}{A flag for if the concepts should include all descendant concepts when imported (modifiable in ATLAS)}

\item{Exclude}{A flag for if the concepts should be excluded when imported (modifiable in ATLAS)}

\item{Atlas}{The WebAPI url for the ATLAS instance}

\item{authMethod}{The authentication method for the ATLAS instance}

\item{Username}{The Username to upload the concept sets to ATLAS under}

\item{Password}{The Password to authorise the user to upload to ATLAS}
}
\description{
Concept pumping function
}
