\name{falconOutput}
\alias{falconOutput}
\docType{data}
\title{
Data frame of CNV calls from falcon
}
\description{
Data frame of CNV calls from falcon.  As calculated in MARATHON notebook section 4.3.2. "Running FALCON for allele-specific copy number profiling."
}
\usage{data(falconOutput)}
\value{
CNV calls from falcon with chromosome, location, and number of exons covered
}
\author{
Hao Chen \email{hxchen@ucdavis.edu}

Gene Urrutia \email{urrutia@live.unc.edu}
}
\examples{
falconOutput
}
\keyword{datasets}