\name{genotype.demo}
\alias{genotype.demo}
\docType{data}
\title{
Demo genotype data for falonx of 39 paired tumor-normal data from breast cancer patients.
}
\description{
Demo genotype data for falonx of 39 paired tumor-normal data from breast cancer patients.
}
\usage{data(genotype.demo)}
\value{
Demo genotype data for falonx of 39 paired tumor-normal data from breast cancer patients.
}
\author{
Yuchao Jiang \email{yuchaoj@email.unc.edu}
}
\examples{
head(genotype.demo)
}
\keyword{datasets}