% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportConceptSets.R
\name{multiplesheets}
\alias{multiplesheets}
\title{Sheet read - a helper function to ingest codes saved as excel workbooks and prepare them for mapping}
\usage{
multiplesheets(fname)
}
\arguments{
\item{fname}{An object of xlsx type that needs to be read in containing named sheets of codes to be mapped, including which vocabulary its supposed to be}
}
\description{
Sheet read - a helper function to ingest codes saved as excel workbooks and prepare them for mapping
}
