\name{mymatrix.demo}
\alias{mymatrix.demo}
\docType{data}
\title{
Demo SNP info for falonx of 39 paired tumor-normal data from breast cancer patients.
}
\description{
Demo SNP info for falonx of 39 paired tumor-normal data from breast cancer patients.
}
\usage{data(mymatrix.demo)}
\value{
Demo SNP info for falonx of 39 paired tumor-normal data from breast cancer patients.
}
\author{
Yuchao Jiang \email{yuchaoj@email.unc.edu}
}
\examples{
head(mymatrix.demo)
}
\keyword{datasets}