\name{primary.demo}
\alias{primary.demo}
\docType{data}
\title{
Demo data of primary neuroblatoma sample for falcon.
}
\description{
Demo data of primary neuroblatoma sample for falcon.
}
\usage{data(primary.demo)}
\value{
Demo data of primary neuroblatoma sample for falcon.
}
\author{
Yuchao Jiang \email{yuchaoj@email.unc.edu}
}
\examples{
dim(primary.demo) # each row corresponds to a SNP
}
\keyword{datasets}