\name{reads.demo}
\alias{reads.demo}
\docType{data}
\title{
Demo allelic read count data for falonx of 39 paired tumor-normal data from breast cancer patients.
}
\description{
Demo allelic read count data for falonx of 39 paired tumor-normal data from breast cancer patients.
}
\usage{data(reads.demo)}
\value{
Demo allelic read count data for falonx of 39 paired tumor-normal data from breast cancer patients.
}
\author{
Yuchao Jiang \email{yuchaoj@email.unc.edu}
}
\examples{
head(reads.demo)
}
\keyword{datasets}