\name{relapse.demo}
\alias{relapse.demo}
\docType{data}
\title{
Demo data of relapse neuroblatoma sample for falcon.
}
\description{
Demo data of relapse neuroblatoma sample for falcon.
}
\usage{data(relapse.demo)}
\value{
Demo data of relapse neuroblatoma sample for falcon.
}
\author{
Yuchao Jiang \email{yuchaoj@email.unc.edu}
}
\examples{
dim(relapse.demo) # each row corresponds to a SNP
}
\keyword{datasets}