% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportConceptSets.R
\name{translate}
\alias{translate}
\title{Title}
\usage{
translate(
  path = "SomeInputofCodes",
  cleaning = TRUE,
  relationships = "Maps to",
  output = "xlsx",
  label = NULL
)
}
\arguments{
\item{path}{A file path to a csv, xlsx or R file that provides the source vocabulary codes to map. Appropriate method should be selected automatically}

\item{cleaning}{Logical trigger to identify if the input cleaning heuristics are required. Defaults to FALSE if the input R file is used}

\item{relationships}{A string that defines which relationship within the hierarchy to use. Defaults to the most useful, "Maps to"}

\item{output}{A string variable chosen from xlsx, conceptSet or cohort that defines what is produced by the function}

\item{label}{A string to identify the ouptut files}
}
\description{
Title
}
